fix: adjusted issue update file location to src/commands/common/issue-update.ts, revert auto-give and reward part comment-update code.
