+++ src/commands/give-up/index.ts
@@ -2,7 +2,7 @@ import { Context } from "probot";
 import {
   createOrUpdateStatus,
   createOrUpdateNotification,
-} from "../api/issue-update/issue-update";
+} from "../common/issue-update";
 import { GiveUpQuery } from "../../queries/GiveUpQuery";
 
 import ChallengeIssueService from "../../services/challenge-issue";
